requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
discord.py
transformers
textblob
django
matplotlib
seaborn
dj-database-url
djangorestframework
django-extensions
psycopg2-binary
whitenoise
tensorflow
tf-keras
awsebcli